feat: better argument handling for cmds
