+++ lib/cmds.js
@@ -1,23 +1,5 @@
-const cmds = [
-    require('./cmd/link.js'),
-    require('./cmd/copy.js'),
-    require('./cmd/exec.js'),
-    require('./cmd/build.js'),
-    require('./cmd/install.js'),
-    require('./cmd/release.js'),
-]
-
-const cmd_obj = {}
-
-cmds.map(
-    cmd =>
-        (cmd_obj[cmd.id] = {
-            fn: cmd.fn,
-            doc: cmd.doc,
-        })
-)
-
-module.exports = {
-    ...cmd_obj,
-    list: cmds.map(cmd => cmd.id),
-}
+module.exports = require('yargs')
+    .usage('usage: $0 <command> [options]')
+    .commandDir('cmds')
+    .demandCommand()
+    .help()
