+++ lib/cmds/README.md
@@ -1,0 +1,17 @@
+<!-- @format -->
+
+# Commands
+
+
+```
+exports.command = ''
+
+exports.describe = ''
+
+exports.builder = {
+}
+
+exports.handler = function (argv) {
+    const { cwd, is_monorepo, root_dir, tool } = argv
+}
+```
