+++ lib/cmds/link.js
