+++ lib/cmds/publish.js
