+++ package.json
@@ -23,7 +23,8 @@
     "fs-extra": "^7.0.1",
     "semver": "^5.6.0",
     "standard-version": "^4.4.0",
-    "tape": "^4.9.1"
+    "tape": "^4.9.1",
+    "yargs": "^12.0.5"
   },
   "publishConfig": {
     "access": "public"
